--- context/state.go
@@ -1,5 +1,0 @@
-package context
-
-// Holds data generated during runtime for output plugin.
-// TODO: Add fields to store. Fields will be required if/when data buffering is implemented.
-type S3State struct{}
